Adding the error calculation to the trilateration.
The planet must now be known before calculating (for decomposition of the error in horizontal/vertical). But that's not a problem as it's updated before the trilateration anyway.
To improve accuracy, only Range combinations that have an error less than twice the median value are kept, as some can have very large errors and compromise the whole set.
In average, we reject about 1/3 of the combinations.

The error estimate is adjusted so that <Actual error/Estimated error> = 1.
The Actual error is more likely to be bellow the estimated one.
When the Actual error overshoots the Estimated one, the error can however sometimes reach a factor of 10.
But generally, High value of Estimated error => High value of Actual error

Signed-off-by: Samuel Albert <samuel_albert38@hotmail.com>